pyproject.toml
[project]
name = "hexlet-code"
version = "0.1.0"
description = "First version of project."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flake8>=7.1.2",
    "flask>=3.1.0",
    "gunicorn>=23.0.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
    "urllib3>=2.3.0",
    "validators>=0.34.0",
]
